requirements.txt
streamlit==1.32.0
sentence-transformers==2.3.1
chromadb==0.4.22
python-docx==1.1.0